photo api

+++ src/main/java/com/olah/gcloud/backup/server/dynamodb/PhotoDao.java
@@ -8,7 +8,6 @@ import com.amazonaws.services.dynamodbv2.document.*;
 import com.amazonaws.services.dynamodbv2.document.spec.QuerySpec;
 import com.amazonaws.services.dynamodbv2.document.utils.ValueMap;
 import com.amazonaws.services.dynamodbv2.model.*;
-import com.amazonaws.services.dynamodbv2.util.TableUtils;
 import com.olah.gcloud.backup.api.model.Photo;
 
 import java.util.*;
@@ -16,16 +15,12 @@ import java.util.stream.Collectors;
 import java.util.stream.StreamSupport;
 
 public class PhotoDao {
-
     public static final String PHOTO_TABLE_NAME = "DeviceStatus";
 
     public static final String FILE_NAME = "FILE_NAME";
     public static final String FOLDER_PATH = "FOLDER_PATH";
-    public static final String STATUS = "STATUS";
+    public static final String STATUS = "STATUS_ATTRIBUTE";
 
-
-    public static final String FOLDER_NAME_WITH_FILE_NAME = "FOLDER_NAME_WITH_FILE_NAME";
-    public static final String FOLDER_NAME_WITH_STATUS = "FOLDER_NAME_WITH_STATUS";
     public static final String FOLDER_STATUS_INDEX = "FOLDER_STATUS_INDEX";
 
     private AmazonDynamoDB client;
@@ -45,7 +40,6 @@ public class PhotoDao {
                 .build();
 
         dynamoDB = new DynamoDB(client);
-
     }
 
 
@@ -53,23 +47,28 @@ public class PhotoDao {
         CreateTableRequest request = new CreateTableRequest();
 
         request.setTableName(PHOTO_TABLE_NAME);
-        request.withKeySchema(new KeySchemaElement(FOLDER_NAME_WITH_FILE_NAME, KeyType.HASH));
+        request.withKeySchema(
+                new KeySchemaElement(FOLDER_PATH, KeyType.HASH),
+                new KeySchemaElement(FILE_NAME, KeyType.RANGE)
+                );
         request.withAttributeDefinitions(
-                new AttributeDefinition(FOLDER_NAME_WITH_FILE_NAME, ScalarAttributeType.S),
-                new AttributeDefinition(FOLDER_NAME_WITH_STATUS, ScalarAttributeType.S)
-        );
+                new AttributeDefinition(FOLDER_PATH, ScalarAttributeType.S),
+                new AttributeDefinition(FILE_NAME, ScalarAttributeType.S),
+                new AttributeDefinition(STATUS, ScalarAttributeType.S)
+                );
         request.withProvisionedThroughput(
                 new ProvisionedThroughput()
                         .withReadCapacityUnits(3L)
                         .withWriteCapacityUnits(3L)
         );
 
-
         GlobalSecondaryIndex globalSecondaryIndex = new GlobalSecondaryIndex();
         globalSecondaryIndex.setIndexName(FOLDER_STATUS_INDEX);
-        globalSecondaryIndex.withKeySchema(new KeySchemaElement(FOLDER_NAME_WITH_STATUS, KeyType.HASH));
+        globalSecondaryIndex.withKeySchema(
+                new KeySchemaElement(FOLDER_PATH, KeyType.HASH), // partition
+                new KeySchemaElement(STATUS, KeyType.RANGE)); //sort
         globalSecondaryIndex.withProjection(new Projection().withProjectionType(ProjectionType.ALL));
-        globalSecondaryIndex.withProvisionedThroughput(                new ProvisionedThroughput()
+        globalSecondaryIndex.withProvisionedThroughput(new ProvisionedThroughput()
                 .withReadCapacityUnits(1L)
                 .withWriteCapacityUnits(1L));
         request.withGlobalSecondaryIndexes(globalSecondaryIndex);
@@ -85,9 +84,10 @@ public class PhotoDao {
 
 
     public void storeOrUpdatePhoto(Photo photo) {
+        System.out.println("Storing photo...");
+        long start = System.currentTimeMillis();
+
         Map<String, AttributeValue> items = new HashMap<>();
-        items.put(FOLDER_NAME_WITH_FILE_NAME, new AttributeValue().withS(createFilePath(photo.getFolderPath(), photo.getFileName()))); // key
-        items.put(FOLDER_NAME_WITH_STATUS, new AttributeValue().withS(createFolderWithStatus(photo.getFolderPath(), photo.getStatus()))); // index key
 
         items.put(FILE_NAME, new AttributeValue().withS(photo.getFileName()));
         items.put(FOLDER_PATH, new AttributeValue().withS(photo.getFolderPath()));
@@ -95,26 +95,22 @@ public class PhotoDao {
 
         PutItemRequest putItemRequest = new PutItemRequest(PHOTO_TABLE_NAME, items);
         client.putItem(putItemRequest);
-    }
 
-    private String createFolderWithStatus(String folderPath, String status) {
-        return folderPath + "_" + status;
-    }
-
-    private String createFilePath(String folderPath, String filename) {
-        return folderPath + "/" + filename;
+        System.out.println("Storing photo took:" + String.valueOf(System.currentTimeMillis() - start));
     }
 
     public Photo getPhoto(String folderPath, String fileName) {
-        String key = createFilePath(folderPath, fileName);
+        long start = System.currentTimeMillis();
 
         Map<String, AttributeValue> keys = new HashMap<>();
-        keys.put(FOLDER_NAME_WITH_FILE_NAME, new AttributeValue().withS(key));
+        keys.put(FILE_NAME, new AttributeValue().withS(fileName));
+        keys.put(FOLDER_PATH, new AttributeValue().withS(folderPath));
+
         GetItemResult getItemResult = client.getItem(PHOTO_TABLE_NAME, keys);
 
         Map<String, AttributeValue> item = getItemResult.getItem();
         if (item == null) {
-            throw new IllegalArgumentException("Could not found photo with key " + key);
+            throw new IllegalArgumentException("Could not found photo with key " + folderPath + "." + fileName);
         }
 
         Photo result = new Photo();
@@ -123,18 +119,21 @@ public class PhotoDao {
         result.setFolderPath(item.get(FOLDER_PATH).getS());
         result.setStatus(Photo.StatusEnum.fromValue(item.get(STATUS).getS()));
 
+        System.out.println("Get photo took:" + String.valueOf(System.currentTimeMillis() - start));
+
         return result;
     }
 
     public Set<Photo> getPhotosByFolderAndStatus(String folderPath, Photo.StatusEnum status) {
+        long start = System.currentTimeMillis();
 
-        String key = createFolderWithStatus(folderPath, status.value());
         Table table = dynamoDB.getTable(PHOTO_TABLE_NAME);
         Index index = table.getIndex(FOLDER_STATUS_INDEX);
 
         QuerySpec spec = new QuerySpec()
-                .withKeyConditionExpression("FOLDER_NAME_WITH_STATUS = :v_key")
-                .withValueMap(new ValueMap().withString(":v_key", key));
+                .withKeyConditionExpression(FOLDER_PATH +" = :v_folder_path and " + STATUS + " = :v_status")
+                .withValueMap(new ValueMap().withString(":v_folder_path", folderPath)
+                        .withString(":v_status", status.value()));
 
         ItemCollection<QueryOutcome> items = index.query(spec);
         Iterator<Item> iterator = items.iterator();
@@ -143,13 +142,18 @@ public class PhotoDao {
 
         java.util.stream.Stream<Item> targetStream = StreamSupport.stream(iterable.spliterator(), false);
 
-        return targetStream.map( item -> {
+        Set<Photo> result = targetStream.map(item -> {
             Photo photo = new Photo();
             photo.setFileName(item.getString(FILE_NAME));
             photo.setFolderPath(item.getString(FOLDER_PATH));
             photo.setStatus(Photo.StatusEnum.fromValue(item.getString(STATUS)));
             return photo;
         }).collect(Collectors.toSet());
+
+
+        System.out.println("getPhotosByFolderAndStatus took :" + String.valueOf(System.currentTimeMillis() - start));
+
+        return result;
     }
 
 
@@ -157,15 +161,22 @@ public class PhotoDao {
     void deleteTable() {
         System.out.println("Deleting table " + PHOTO_TABLE_NAME + "...");
 
+        boolean tableAlreadyExists =
+        StreamSupport.stream(dynamoDB.listTables().spliterator(), false).anyMatch(table -> table.getTableName().equals(PHOTO_TABLE_NAME));
+
+        if(!tableAlreadyExists) {
+            return;
+        }
+
         Table table = dynamoDB.getTable(PHOTO_TABLE_NAME);
+
         table.delete();
 
         // Wait for table to be deleted
         System.out.println("Waiting for " + PHOTO_TABLE_NAME + " to be deleted...");
         try {
             table.waitForDelete();
-        }
-        catch (InterruptedException e) {
+        } catch (InterruptedException e) {
             e.printStackTrace();
         }
     }
